added day count, fixed specialPost bug, added controllers in differant files
