+++ server/models/Post.js
@@ -4,14 +4,17 @@ const PostSchema = new Mongoose.Schema({
     body: {
         type: String
     },
-    author:{
-        type:Mongoose.Schema.Types.ObjectId,
-        ref:'User'
+    author: {
+        type: Mongoose.Schema.Types.ObjectId,
+        ref: 'User'
     },
-    challenge:{
+    challenge: {
         type: Number,
     },
-    isSpecialPost: Boolean
+    isSpecialPost: Boolean,
+    day: {
+        type: Number
+    }
 }, { timestamps: true })
 
 module.exports = Mongoose.model('Post', PostSchema, 'Posts')
